Fix broken wheel listeners
Chrome seems to have made a change overnight that blocks JS from running
if they detect non-passive event listeners that mess with scrolling
https://www.chromestatus.com/feature/5745543795965952

This wasn't an issue as of yesterday, but today the app doesn't work.

Preventing the event's default behavior is unecessary in this app
because the CSS has resized everything to fit in the window, thus
there is nothing to scroll.